backend/pyproject.toml
[project]
name = "project_name_back"
version = "0.1.0"
description = "API Template for the ENSAI 2nd year IT project"
authors = [{ name = "Lastname Firstname" }, { name = "Author2" }]
requires-python = ">=3.13"

dependencies = [
    "fastapi",
    "inquirerPy",
    "psycopg2-binary",
    "python-dotenv",
    "PyYAML",
    "requests",
    "regex",
    "uvicorn",
]

# dependencies used only for development, not for runtime
[project.optional-dependencies]
dev = ["bandit", "coverage", "pylint", "pytest", "ruff"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" } #source files are in folder src

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]                        # si tu veux ignorer la longueur de ligne

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["*test*.py"]
addopts = "-v -s"

[tool.coverage.run]
branch = true
include = ["src/service/*", "src/dao/*"]
omit = ["*/__init__.py"]

[tool.coverage.report]
show_missing = true

[tool.coverage.html]
directory = "coverage_report"
